requirements.txt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
python-dateutil
beautifulsoup4
openai
pydantic
pytz
lxml
python-dotenv